优化了文件缓存

+++ src/Cache/FileCache.php
@@ -4,6 +4,13 @@ namespace Sapi\Cache;
 
 class FileCache implements Cache
 {
+    private string $cacheFolderPath;
+
+    public function __construct()
+    {
+        $this->cacheFolderPath = $this->createCacheFileFolder();
+    }
+
     public function set(string $key, string $value, $expire = 600): bool
     {
         $filePath = $this->generateFilePath($key);
@@ -17,11 +24,19 @@ class FileCache implements Cache
         return file_put_contents($filePath, serialize($param)) !== false;
     }
 
-
     public function get(string $key): string
     {
         $filePath = $this->generateFilePath($key);
+
+        if (!file_exists($filePath)) {
+            return '';
+        }
+
         $data = file_get_contents($filePath);
+        if ($data === false) {
+            return '';
+        }
+
         $param = unserialize($data);
         if (time() > ($param['create_tm'] + $param['expire'])) {//超时
             $this->delete($key);
@@ -29,7 +44,6 @@ class FileCache implements Cache
         }
         return $param['data'];
     }
-
 
     public function delete(string $key)
     {
@@ -39,25 +53,30 @@ class FileCache implements Cache
 
         $filePath = $this->generateFilePath($key);
 
-        @unlink($filePath);
+        if (file_exists($filePath)) {
+            unlink($filePath);
+        }
     }
-
 
     private function generateFilePath(string $key): string
     {
-        $directoryPath = ROOT_PATH . "storage/cache/" . $this->hashToMod($key);
+        $directoryPath = $this->cacheFolderPath . $this->hashToMod($key);
 
         if (!is_dir($directoryPath)) {
             mkdir($directoryPath, 0755, true);
         }
-        return $directoryPath . '/' . sha1($key);
+
+        return $directoryPath . DIRECTORY_SEPARATOR . sha1($key);
+    }
+
+    private function createCacheFileFolder(): string
+    {
+        return ROOT_PATH . "storage" . DIRECTORY_SEPARATOR . "cache" . DIRECTORY_SEPARATOR;
     }
 
     private function hashToMod(string $key, $numShards = 100): string
     {
-        $hashValue = md5($key); // 使用 md5() 生成哈希值
-        // 取模运算
+        $hashValue = md5($key);
         return strval((hexdec(substr($hashValue, -4)) % $numShards));
     }
-
 }
